\name{runResilate}
\alias{runResilate}
\title{ Start a Menu of Models for Resilation }
\description{
  Start a GUI that controls which models to pass into the 
  \code{resilate} function.
}
\usage{
runResilate()
}
\details{
  Looks at the names of R-code (\code{*.r}) in the folder
  \samp{.../PBSresilate/examples} and uses the prefixes as 
  available models.
}
\value{
  No value returned.
}
\author{
  \href{mailto:rowan.haigh@dfo-mpo.gc.ca}{Rowan Haigh}, Program Head -- Offshore Rockfish\cr
  Pacific Biological Station (PBS), Fisheries & Oceans Canada (DFO), Nanaimo BC\cr
  \emph{opus locus}: Institute of Ocean Sciences (IOS), Sidney BC\cr
  Last modified \code{Rd: 2018-08-21}
}
\seealso{ 
	\code{\link{resilate}}
}
\keyword{ hplot }
