\name{runResilate}
\alias{runResilate}
\title{ Start a Menu of Models for Resilation }
\description{
  Start a GUI that controls which models to pass into the 
  \code{resilate} function.
}
\usage{
runResilate()
}
\details{
  Looks at the names of R-code (\code{*.r}) in the folder
  \samp{.../PBSresilate/examples} and uses the prefixes as 
  available models.
}
\value{
  No value returned.
}
\author{
  Rowan Haigh, Pacific Biological Station, Nanaimo BC 
}
\seealso{ 
	\code{\link{resilate}}
}
\keyword{ hplot }
